refactor: update import paths for consistency and improve component structure
