+++ src/components/detailpage/index.ts
@@ -1,3 +1,2 @@
-// export Homepage Components
 export { default as Related } from "./ProductRelated.tsx";
 export { default as ProductInfo } from "./ProductInfo.tsx";
