+++ src/config/index.ts
@@ -1,0 +1,2 @@
+export { headerLinks, footerLinks } from "./links.ts";
+export { typographySizes } from "./typographySizes.ts";
